+++ src/ogg/flac/read.rs
@@ -2,11 +2,15 @@ use super::block::Block;
 use super::FlacFile;
 use crate::error::{FileDecodingError, Result};
 use crate::file::FileType;
-#[cfg(feature = "vorbis_comments")]
-use crate::ogg::{read::read_comments, tag::VorbisComments};
-#[cfg(feature = "vorbis_comments")]
-use crate::picture::Picture;
+use crate::id3::ID3FindResults;
+#[cfg(feature = "id3v2")]
+use crate::id3::{find_id3v2, v2::read::parse_id3v2};
 use crate::properties::FileProperties;
+#[cfg(feature = "vorbis_comments")]
+use crate::{
+	ogg::{read::read_comments, tag::VorbisComments},
+	picture::Picture,
+};
 
 use std::io::{Read, Seek, SeekFrom};
 
@@ -40,6 +44,25 @@ pub(super) fn read_from<R>(data: &mut R, read_properties: bool) -> Result<FlacFi
 where
 	R: Read + Seek,
 {
+	let mut flac_file = FlacFile {
+		#[cfg(feature = "id3v2")]
+		id3v2_tag: None,
+		#[cfg(feature = "vorbis_comments")]
+		vorbis_comments: None,
+		properties: FileProperties::default(),
+	};
+
+	// It is possible for a FLAC file to contain an ID3v2 tag
+	if let ID3FindResults(Some(header), Some(content)) = find_id3v2(data, true)? {
+		#[cfg(feature = "id3v2")]
+		{
+			let reader = &mut &*content;
+
+			let id3v2 = parse_id3v2(reader, header)?;
+			flac_file.id3v2_tag = Some(id3v2)
+		}
+	}
+
 	let stream_info = verify_flac(data)?;
 	let stream_info_len = (stream_info.end - stream_info.start) as u32;
 
@@ -75,6 +98,12 @@ where
 		}
 	}
 
+	#[cfg(feature = "vorbis_comments")]
+	{
+		flac_file.vorbis_comments =
+			(!(tag.items.is_empty() && tag.pictures.is_empty())).then(|| tag);
+	}
+
 	let (stream_length, file_length) = {
 		let current = data.seek(SeekFrom::Current(0))?;
 		let end = data.seek(SeekFrom::End(0))?;
@@ -82,15 +111,11 @@ where
 		(end - current, end)
 	};
 
-	let properties = if read_properties {
+	flac_file.properties = if read_properties {
 		super::properties::read_properties(&mut &*stream_info.content, stream_length, file_length)?
 	} else {
 		FileProperties::default()
 	};
 
-	Ok(FlacFile {
-		properties,
-		#[cfg(feature = "vorbis_comments")]
-		vorbis_comments: (!(tag.items.is_empty() && tag.pictures.is_empty())).then(|| tag),
-	})
+	Ok(flac_file)
 }